pyproject.toml
[tool.poetry]
name = "placebo25"
version = "0.1.0"
description = "Test task for Placebo/25"
authors = ["Алина Воронцова <vorontsova-alina@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
black = "^23.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
